Data-Collection/requirements.txt
tweepy==3.6.0
kafka-python==1.4.2
